+++ index.js
@@ -1,10 +1,22 @@
 'use strict';
 
+const fs = require('fs-extra');
 const path = require('path');
-const config = require(path.join(process.cwd(), 'skyuxconfig.json'));
+
+function getConfig() {
+  let config = {};
+  const configPath = path.join(process.cwd(), `skyuxconfig.json`);
+
+  if (fs.existsSync(configPath)) {
+    config = fs.readJsonSync(configPath);
+  }
+
+  return config;
+}
 
 module.exports = {
   runCommand: (command, argv) => {
+    const config = getConfig();
     switch (command) {
       case 'stache-add-search-spec':
         require('./src/add-search-spec')(argv, config);
@@ -27,9 +39,12 @@ module.exports = {
       case 'stache-remove-search-spec':
         require('./src/remove-search-spec')(argv, config);
         break;
+      case 'version':
+        require('./src/version')();
+        break;
       default:
         return false;
-      }
+    }
     return true;
   }
 };